BasicGeminiVisionApp/requirements.txt
streamlit
google_generativeai
python_dotenv
Pillow